backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt<4.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.127.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest-cov>=7.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy[asyncio]>=2.0.45",
    "uvicorn>=0.40.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "cloudinary>=1.44.1",
]
